major checkin for sweden demo I got the loan invoice working (uses mulitple file subreports) cj did the report (thx) I reworked some of the notification stuff for tasks
git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@1105 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39

+++ src/edu/ku/brc/dbsupport/DBTableIdMgr.java
@@ -308,7 +308,7 @@ public class DBTableIdMgr
 	 * @param recordId a single Record Id
 	 * @return a query object
 	 */
-	public static String getQueryForTable(final int tableId, final int recordId)
+	public static String getQueryForTable(final int tableId, final long recordId)
 	{
 		TableInfo tableInfo = instance.hash.get(tableId);
 		if (tableInfo != null)
